requirements.txt
flask==3.1.2
gunicorn==23.0.0
PyMuPDF==1.26.5
sentence-transformers==5.1.2
scikit-learn==1.7.2
numpy
pandas
torch
transformers
python-dotenv
requests
tqdm
SQLAlchemy>=2.0.0
pymysql
openpyxl
cryptography
